Give a basic style to the book card container
Basic styling was given to the book card container in order to test
the addition of a few test books in the user's library.

There was also some beginning code added to the javascript file that
will become the building blocks of implementing the functionality that
will allow the user to add books with their own inputs.