Remove the dom element creation from the script
Changed the implementation of displaying the new book form from
javascript dom element creation to having the form element pre-written
in the html code.

This should give better performance since the dom elements don't need
to be created every time the form needs to be displayed. All that needs
to be done is hide or show the display of the new book button and the
form element.