Implement the form submission functionality
The "new book" form will now take the user input data and store the
form data into a FormData object that holds all the data that the user
input.

Once the form is submitted, the form inputs are cleared with the
.reset() method so the user can have a clean form the next time they
need to add a book.

The new book button is displayed and the form is hidden when the user
submits the book as well.

The addBookToLibrary function was also implemented to include as an
operation when the user submits the form. The function adds a new
Book object to the myLibrary list whenever the user submits a new book
form.