Enable displaying of books from the library array
Books that were added to the myLibrary array via the new book form
are now displayed in the book grid.

A reloadLibraryDisplay() function has been implemented because I
believe it's simplest to just have it be called whenever the myLibrary
array is altered. For now, the only time it's called is when a book is
removed from the array or when a new book is added.

Whenever reloadLibraryDisplay() is called, the child elements of the
book grid element need to be cleared and children need to be
re-appended for each book in the library.

createBookCardDOMElement() was implemented to make sure a DOM
element is properly created for each book card element that needs to be
displayed in the book grid. This function is called cleanly within
reloadLibraryDisplay() for every book in the library.